feat : 노선 중간역 제거 기능 구현

+++ src/main/java/nextstep/subway/line/Line.java
@@ -78,8 +78,11 @@ public class Line {
         newSection.registerLine(this);
     }
 
-    public void deleteDownSection(Station deleteStation) {
-        sections.deleteDownStation(deleteStation);
+    public void deleteSection(Station deleteStation) {
+        if(sections.isLastSection(deleteStation)) {
+            sections.deleteDownStation(deleteStation);
+        }
+        sections.deleteMiddleStation(deleteStation);
     }
 
     @Override
